Update ISIDASeq.py
The normalization procedure for the simple k-mers was modified. The normalization block for the position-centered k-mers remains to be modified in the same manner as for the simple k-mers. An option allowing to work directly with contigs was added (the contig file is required to have distinct line for each of the genome sequences, whereas for each of the genome sequences represented in the form of contigs, the contigs must be separated by comma, e.g. actgaattc,attgcaa,atcggatc). The k-mers counts for each of the contigs are calculated separately and then divided by the total number of k-mers counts in the whole sequence to yield normalized k-mers.